# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gbolix API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List current user's projects
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project (service request)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project (client)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  # ── Files ─────────────────────────────────────────────────────────────────────
...
  /files/upload:
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a file (base64 JSON body)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileUploadInput"
      responses:
        "201":
          description: Uploaded file record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRecord"
  /files/download/{filename}:
    get:
      operationId: downloadFile
      tags: [files]
      summary: Download a file by stored filename
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File binary
  /files/{id}:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Messages ──────────────────────────────────────────────────────────────────
  /projects/{projectId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /projects/{projectId}/messages/read:
    post:
      operationId: markMessagesRead
      tags: [messages]
      summary: Mark all messages in a project as read
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
                required:
                  - updated

  # ── Notifications ─────────────────────────────────────────────────────────────
...
  /admin/projects:
    get:
      operationId: adminListProjects
      tags: [admin]
      summary: List all projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminProject"
  /admin/projects/{id}:
    patch:
      operationId: adminUpdateProject
      tags: [admin]
      summary: Update a project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProject"
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        serviceType:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [submitted, queued, in_progress, review, completed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        price:
          type: ["number", "null"]
        hasConversation:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - serviceType
        - title
        - description
        - status
        - priority
        - hasConversation
        - createdAt
        - updatedAt
    ProjectInput:
      type: object
      properties:
        serviceType:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        price:
          type: number
      required:
        - serviceType
        - title
        - description
        - priority
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
    FileRecord:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        projectId:
          type: ["integer", "null"]
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        url:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - filename
        - originalName
        - mimeType
        - size
        - url
        - createdAt
    FileUploadInput:
      type: object
      properties:
        filename:
          type: string
        mimeType:
          type: string
        data:
          type: string
        projectId:
          type: integer
      required:
        - filename
        - mimeType
        - data
    Message:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        senderRole:
          type: string
          enum: [owner, admin, freelancer, client]
        content:
          type: string
        fileUrl:
          type: ["string", "null"]
        fileName:
          type: ["string", "null"]
        fileMimeType:
          type: ["string", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - projectId
        - senderId
        - senderName
        - senderRole
        - content
        - isRead
        - createdAt
    MessageInput:
      type: object
      properties:
        content:
          type: string
        fileData:
          type: string
        fileName:
          type: string
        fileMimeType:
          type: string
...
    AdminProject:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        serviceType:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [submitted, queued, in_progress, review, completed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        price:
          type: ["number", "null"]
        internalNotes:
          type: ["string", "null"]
        hasConversation:
          type: boolean
        clientName:
          type: string
        clientEmail:
          type: string
        assignedFreelancers:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
            required:
              - id
              - name
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - serviceType
        - description
        - status
        - priority
        - hasConversation
        - clientName
        - clientEmail
        - assignedFreelancers
        - createdAt
        - updatedAt
    AdminProjectUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [submitted, queued, in_progress, review, completed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        internalNotes:
          type: string
        price:
          type: number
...